Changed inventory mapping id from int to long

+++ green-river/src/main/scala/consumer/elastic/mappings/admin/InventorySearchView.scala
@@ -9,15 +9,15 @@ import consumer.elastic.mappings._
 final case class InventorySearchView()(implicit ec: EC) extends AvroTransformer {
   def topic() = "inventory_search_view"
   def mapping() = esMapping(topic()).fields(
-      field("id", IntegerType),
+      field("id", LongType),
       field("sku", StringType).analyzer("upper_cased"),
       field("stockItem").nested(
-          field("id", IntegerType),
+          field("id", LongType),
           field("sku", StringType).analyzer("upper_cased"),
           field("defaultUnitCost", IntegerType)
       ),
       field("stockLocation").nested(
-          field("id", IntegerType),
+          field("id", LongType),
           field("name", StringType).analyzer("autocomplete"),
           field("type", StringType).index("not_analyzed")
       ),
